+++ pong_front/src/components/auth/Login/LoginButton.style.ts
@@ -1,8 +1,6 @@
 import { styled } from "@mui/material"
-import { Button } from "@mui/material"
-import { theme } from "../../../common/styles/Theme.style"
 
-import { DefaultButton } from "../../../common/styles/DefaultButton"
+import { DefaultButton } from "../../common"
 
 export const LoginButton = styled(DefaultButton) `
     font-size: 2rem;
